% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{alt_check}
\alias{alt_check}
\title{Suggest Alt Tag Oddities}
\usage{
alt_check(attr_df)
}
\arguments{
\item{attr_df}{A data.frame with image attributes produced by alt_get_attr()}
}
\value{
A data.frame object.
}
\description{
Infer whether the alt tag is missing or strange.
}
