% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{alt_get}
\alias{alt_get}
\title{Get HTML Attributes from the Image Elements of a Web Page}
\usage{
alt_get(url, all_attributes = FALSE)
}
\arguments{
\item{url}{A character string describing the URL of a webpage.}

\item{all_attributes}{Logical. Do you want to return all the image attributes
(\code{TRUE}), or just the \code{src} (image source) and \code{alt}
(alt text) attributes (\code{FALSE})?}
}
\value{
A tibble object with classes "tbl_df", "tbl", "data.frame").
}
\description{
Scrape a web page's HTML and extract into a tibble the attributes
    from each image element. Only collects attributes if they exist.
}
