% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{alt_get_attr}
\alias{alt_get_attr}
\title{Get the Image Attributes for a Given Page}
\usage{
alt_get_attr(url)
}
\arguments{
\item{url}{A string describing the URL of a webpage.}
}
\value{
A data.frame object.
}
\description{
Scrapes a web page for the image tag and extracts the tags into a data frame.
}
