% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altcheckr-package.R
\docType{package}
\name{altcheckr-package}
\alias{altcheckr}
\alias{altcheckr-package}
\title{altcheckr: Assess Image Alt Text on Websites}
\description{
Inspect a webpage for the presence of, and problems with, image alt text.
    Alt text provides a description of image content for people who use assistive
    technologies like screen readers. It's encoded in the 'alt' attribute of the 
    'img' HTML element.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/matt-dray/altcheckr/}
  \item \url{https://matt-dray.github.io/altcheckr/}
  \item Report bugs at \url{https://github.com/matt-dray/altcheckr/issues}
}

}
\author{
\strong{Maintainer}: Matt Dray \email{mwdray@gmail.com}

}
\keyword{internal}
