% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_get}
\alias{example_get}
\title{Example of alt_get() Output}
\format{
A data frame with 11 rows and 2 variables:
\describe{
  \item{src}{a fake image source}
  \item{alt}{dummy alt text with different types of problems}
  ...
}
}
\usage{
example_get
}
\description{
A dummy data frame to mimic output from \code{\link{alt_get}}.
}
\keyword{datasets}
