% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{alt_check}
\alias{alt_check}
\title{Check for Problems with the Alt Text for Images on a Web Page}
\usage{
alt_check(attributes_df, max_char = 200, min_char = 50,
  file_ext = ".jpg$|.jpeg$|.png$|.svg$|.gif$",
  redundant_pattern = "image|picture|photo|graph|plot|diagram")
}
\arguments{
\item{attributes_df}{A data.frame/tibble with image attributes, produced by
\code{altcheckr::alt_get()}.}

\item{max_char}{A numeric value. If the number of characters in the alt
text is larger than this then it's flagged as being long}

\item{min_char}{A numeric value. If the number of characters in the alt
text is smaller than this then it's flagged as being short.}

\item{file_ext}{A character string. Regular expression of image file
extensions that might signify a file name has been used for the alt text.}

\item{redundant_pattern}{A character string. Regular expression for phrase(s)
to flag as being redundant in alt text ("A picture of" isn't necessary,
for example).}
}
\value{
A tibble object with classes "tbl_df", "tbl", "data.frame".
}
\description{
Infer whether an image's alt tag is missing or could be improved.
}
