% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{alt_check}
\alias{alt_check}
\title{Check for Problems with the Alt Text for Images on a Web Page}
\usage{
alt_check(
  attributes_df,
  max_char = 125,
  min_char = 20,
  file_ext = ".jpg$|.jpeg$|.png$|.svg$|.gif$",
  redundant_pattern = "image|picture|photo|graph|plot|diagram"
)
}
\arguments{
\item{attributes_df}{A data.frame/tibble with image attributes, produced by
\code{\link{alt_get}}.}

\item{max_char}{A numeric value. Alt text longer than this is flagged.}

\item{min_char}{A numeric value. Alt text shorter than this is flagged.}

\item{file_ext}{A character string. A regular expression of image file
extensions.}

\item{redundant_pattern}{A character string. A regular expression of
'redundant' phrases in alt text.}
}
\value{
A tibble object with classes \code{tbl_df}, \code{tbl} and
    \code{data.frame}. In addition to columns provided by
    \code{\link{alt_get}}, also returns:
    \itemize{
        \item \code{alt_exists} "Exists", "Missing" or intentionally "Empty".
        \item \code{nchar_count} Integer. Number of characters.
        \item \code{nchar_assess} "Short", "Long" or "OK", depending on inputs
            to \code{min_char} and \code{max_char}.
        \item \code{file_ext} Logical. Does it look like the alt text might
            just be a filename (e.g. ends with ".jpg")?
        \item \code{self_evident} Logical. Is a redundant phrase used in the
            alt text (e.g. "a picture of")?
        \item \code{terminal_punct} Logical. Does the alt text end with
            terminal punctuation, like a period, to allow a screen reader to
            parse it as a sentence?
        \item \code{spellcheck} Character. Words to check for misspelling.
            These could be misread by a screenreader.
        \item \code{not_basic} Character list column. Words that match to
            the 850 words of Charles Kay Ogden's Basic English.
     }
}
\description{
Infer whether an image's alt tag is missing or could be improved.
}
\examples{
\dontrun{
    url <- "https://alphagov.github.io/accessibility-tool-audit/test-cases.html#images"
    attr_df <- alt_get(url)
    alt_check(attr_df)
    }
}
