% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{alt_check}
\alias{alt_check}
\title{Check for Image Alt Tag Problems}
\usage{
alt_check(attr_df)
}
\arguments{
\item{attr_df}{A data.frame/tibble with image attributes produced by \code{altcheckr::alt_get_attr()}.}
}
\value{
A data.frame/tibble object.
}
\description{
Infer whether an image's alt tag is missing or could be improved.
}
