% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{alt_get_attr}
\alias{alt_get_attr}
\title{Get a Web Page's Image Attributes}
\usage{
alt_get_attr(url)
}
\arguments{
\item{url}{A string describing the URL of a webpage.}
}
\value{
A data.frame object.
}
\description{
Scrapes a web page and extracts the image tags into a data frame.
}
