% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altcheckr-package.R
\docType{package}
\name{altcheckr-package}
\alias{altcheckr}
\alias{altcheckr-package}
\title{altcheckr: Check Image Alt Text on Websites}
\description{
Inspect a webpage for problems with the accessibility of image alt text.
}
\author{
\strong{Maintainer}: Matt Dray \email{mwdray@gmail.com}

}
\keyword{internal}
