% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altcheckr-package.R
\docType{package}
\name{altcheckr-package}
\alias{altcheckr}
\alias{altcheckr-package}
\title{altcheckr: Assess Image Alt Text on a Web Page}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Scrape image element attributes from a webpage, detect alternative
    (alt) text and assess it with simple heuristics. Alt text is important for users
    of assistive technologies, like screen readers, for understanding the content
    of images. This package should be used in conjunction with other accessibility
    assessment tools for more comprehensive coverage.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/matt-dray/altcheckr/}
  \item \url{https://matt-dray.github.io/altcheckr/}
  \item Report bugs at \url{https://github.com/matt-dray/altcheckr/issues}
}

}
\author{
\strong{Maintainer}: Matt Dray \email{mwdray@gmail.com}

}
\keyword{internal}
